Extract gravity tests from library
The gravity test cases are fairly involved integration tests. Therefore they
were moved out of tree as integration tests. This keeps the code base a bit
more organized.
